feat(schema-engine-wasm): wasm-compatible sql-schema-describer
