+++ schema-engine/sql-introspection-tests/Cargo.toml
@@ -6,7 +6,9 @@ edition = "2021"
 [dependencies]
 schema-connector = { path = "../connectors/schema-connector" }
 sql-schema-connector = { path = "../connectors/sql-schema-connector" }
-sql-schema-describer = { path = "../sql-schema-describer" }
+sql-schema-describer = { path = "../sql-schema-describer", features = [
+    "all-native",
+] }
 psl = { workspace = true, features = ["all"] }
 test-macros = { path = "../../libs/test-macros" }
 user-facing-errors = { path = "../../libs/user-facing-errors", features = [
